Add functionality to bold substrs in passages
I've written one set of functions to do this in a tremendously pain
in the ass way, and other just uses a simple regex. I've no idea which
is faster, and I've also got no idea how to measure that at this moment
but I'll just leave them as is for now.